% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_input.R
\name{genericInputData}
\alias{genericInputData}
\title{Make basic input data that can be used across models}
\usage{
genericInputData(grid_width)
}
\arguments{
\item{grid_width}{Number of cells in X direction (assuming X = Y for square grid)}
}
\value{
An object of class `list` containing adjacency matrix information, etc
}
\description{
Make basic input data that can be used across models
}
