% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_input.R
\name{makeMooreNeighbors}
\alias{makeMooreNeighbors}
\title{Make data frame with moore neighbors}
\usage{
makeMooreNeighbors(coords)
}
\arguments{
\item{grid_width}{Number of cells in X direction (assuming X = Y for square grid)}
}
\value{
An object of class `list` containing adjacency matrix information, etc
}
\description{
Make data frame with moore neighbors
}
