% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_input.R
\name{plotMap}
\alias{plotMap}
\title{Sample map with fixed proportions in each cell}
\usage{
plotMap(sample_df, iter = FALSE)
}
\arguments{
\item{iter}{Which model iteration to use (defaults to random selection)}

\item{Item}{of class `data.frame` with sampled values}
}
\description{
Sample map with fixed proportions in each cell
}
