% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_input.R
\name{sampleProportion}
\alias{sampleProportion}
\title{Sample map with fixed proportions in each cell}
\usage{
sampleProportion(grid_width, i_target = 0, i_sd = 0.1, mean_target = 0.9,
  mean_sd = 0.1, nchain = 1, niter = 1000)
}
\arguments{
\item{grid_width}{X dimension of grid}

\item{i_target}{Value of Moran's I we're shooting for}

\item{i_sd}{Value of SD of Moran's I we're trying to get to}

\item{mean_target}{Mean proportion}

\item{mean_sd}{SD of mean across samples}

\item{nchain}{Number of chains}

\item{niter}{Number of iterations}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Sample map with fixed proportions in each cell
}
