% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_input.R
\name{unpackProportionModel}
\alias{unpackProportionModel}
\title{Unpack fitted model into a more plotting-friendly format}
\usage{
unpackProportionModel(data_in, m)
}
\arguments{
\item{data_in}{Input data for model}

\item{m}{Fitted model of class `stanmodel`}
}
\value{
An object of class `data.frame` with sampled values
}
\description{
Unpack fitted model into a more plotting-friendly format
}
